Add support for default predictions
This branch adds a new json field to the EgaUser model, so preferences can be
stored in a per user basis.

Specifically, the branch allows for a user to set a default prediction, which
could be used when calculating the score for each match, assuming the user
did not set an explicit prediction for that match.

As a side change, a test factory was added in order to allow for easier test
writing, and some black formatting error were corrected.

Lastly, UI work for surfacing this feature will be done in a future iteration.